requirements.txt
defusedxml==0.6.0
PyYAML==5.3
